packages/core/pyproject.toml
[project]
name = "nichefinder-core"
version = "0.1.0"
description = "Core models and services for AI Nichefinder."
requires-python = ">=3.12,<3.15"
dependencies = [
  "beautifulsoup4>=4.13.4",
  "google-genai>=1.13.0",
  "httpx>=0.28.1",
  "jinja2>=3.1.6",
  "langgraph>=1.1.6",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.8.0",
  "playwright>=1.52.0",
  "pytrends>=4.9.2",
  "python-frontmatter>=1.1.0",
  "readability-lxml>=0.8.1",
  "rich>=14.1.0",
  "sqlmodel>=0.0.22",
  "tenacity>=9.1.2",
  "lxml>=5.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nichefinder_core"]
